refactor(allocator): extract VM telemetry routes into routes/vm_telemetry.py
Moves the six client-VM push endpoints and two operator read endpoints
for VM status/logs/metrics out of main.py into a dedicated blueprint,
since both directions operate on the same underlying columns. Hoists
MAX_LOG_SIZE to module scope and switches to the strip_ansi helper.
